+++ contracts/finance/andromeda-splitter/src/testing/tests.rs
@@ -35,7 +35,7 @@ fn init(deps: &mut TestDeps) -> Response {
     let msg = InstantiateMsg {
         owner: Some(OWNER.to_string()),
         kernel_address: MOCK_KERNEL_CONTRACT.to_string(),
-        recipients: mock_recipient,
+        recipients: Some(mock_recipient),
         lock_time: Some(Expiry::FromNow(Milliseconds(86400000))),
         default_recipient: None,
     };
@@ -70,7 +70,7 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![],
+        recipients: Some(vec![]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -87,7 +87,7 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![],
+        recipients: Some(vec![]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -103,10 +103,10 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![AddressPercent {
+        recipients: Some(vec![AddressPercent {
             recipient: Recipient::from_string(some_address.to_string()),
             percent: Decimal::percent(100),
-        }],
+        }]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -122,7 +122,7 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![],
+        recipients: Some(vec![]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -137,7 +137,7 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![],
+        recipients: Some(vec![]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -152,10 +152,10 @@ fn test_different_lock_times() {
     let msg = InstantiateMsg {
         owner: Some(owner.to_string()),
         kernel_address: kernel_address.to_string(),
-        recipients: vec![AddressPercent {
+        recipients: Some(vec![AddressPercent {
             recipient: Recipient::from_string(some_address),
             percent: Decimal::percent(100),
-        }],
+        }]),
         lock_time: Some(lock_time),
         default_recipient: None,
     };
@@ -232,7 +232,7 @@ fn test_execute_update_recipients() {
         },
     ];
     let msg = ExecuteMsg::UpdateRecipients {
-        recipients: duplicate_recipients,
+        recipients: Some(duplicate_recipients),
     };
 
     let info = message_info(&Addr::unchecked(OWNER), &[]);
@@ -252,7 +252,7 @@ fn test_execute_update_recipients() {
         },
     ];
     let msg = ExecuteMsg::UpdateRecipients {
-        recipients: recipients.clone(),
+        recipients: Some(recipients.clone()),
     };
 
     let incorrect_owner = deps.api.addr_make("incorrect_owner");